create a useBaseFetch that useLazyFetch and useFetch use
