+++ src/useFetch.js
@@ -1,24 +1,11 @@
-import { useState, useEffect } from 'react';
-import axios from 'axios';
+import { useEffect } from 'react';
+import useBaseFetch from './useBaseFetch';
 
 export default config => {
-  const [data, setData] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState('');
+  const [getData, { data, error, loading }] = useBaseFetch(config);
 
   useEffect(() => {
-    (async () => {
-      try {
-        setLoading(true);
-        const res = await axios(config);
-        setData(res.data);
-        setLoading(false);
-        setError('');
-      } catch (e) {
-        setError(e);
-        setLoading(false);
-      }
-    })();
+    getData();
   }, [config.url]);
 
   return { data, error, loading };
